% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{add_seq}
\alias{add_seq}
\title{Add a sequence variable within USUBJID}
\usage{
add_seq(df, seqvar)
}
\arguments{
\item{df}{A data.frame containing a USUBJID column}

\item{seqvar}{The name of the sequence variable to create (e.g., "VSSEQ")}
}
\value{
A data.frame with the newly created sequence variable
}
\description{
Generates a sequential integer (as character) grouped by USUBJID, 
commonly used for SDTM --SEQ variables.
}
