% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam_engine.R
\name{build_adam_dataset}
\alias{build_adam_dataset}
\title{Build an ADaM dataset from specifications}
\usage{
build_adam_dataset(spec_yaml_path, source_data)
}
\arguments{
\item{spec_yaml_path}{A character string representing the path to the ADaM YAML spec}

\item{source_data}{A named list of data.frames containing the upstream datasets}
}
\value{
A data.frame containing the fully derived ADaM dataset
}
\description{
Generates an ADaM dataset based on a parsed YAML template, drawing data 
from previously built SDTM or intermediate ADaM datasets. It handles data merges, 
type conversions, variable mapping, derivations, and cuts specified in the YAML.
}
