% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_functions.R
\name{calc_rfendtc}
\alias{calc_rfendtc}
\title{Calculate RFENDTC from RFSTDTC and DSSTDY}
\usage{
calc_rfendtc(rfstdtc, dsstdy, built_domains = NULL, ...)
}
\arguments{
\item{rfstdtc}{A character vector of Reference Start Dates}

\item{dsstdy}{A character or numeric vector of Disposition Study Days}

\item{built_domains}{Optional list of built domains}

\item{...}{Additional arguments}
}
\value{
A character vector of derived Reference End Dates
}
\description{
Derives the Reference End Date (RFENDTC) by adding the Disposition Study Day 
(DSSTDY) minus 1 to the Reference Start Date (RFSTDTC).
}
