% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{epoch_of}
\alias{epoch_of}
\title{Determine EPOCH from visit name}
\usage{
epoch_of(visit)
}
\arguments{
\item{visit}{A character string representing the visit name}
}
\value{
A character string ("SCREENING" or "TREATMENT")
}
\description{
Simple heuristic to determine EPOCH (e.g., SCREENING or TREATMENT) 
based on the visit name.
}
