% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odm_parser.R
\name{extract_metadata_summary}
\alias{extract_metadata_summary}
\title{Extract metadata summary}
\usage{
extract_metadata_summary(df_long)
}
\arguments{
\item{df_long}{Long format dataframe}
}
\value{
Dataframe with deduplicated metadata and sample values
}
\description{
Extract metadata summary
}
