% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs_functions.R
\name{get_bmi}
\alias{get_bmi}
\title{Calculate Body Mass Index (BMI)}
\usage{
get_bmi(height, weight)
}
\arguments{
\item{height}{A numeric vector of heights in centimeters}

\item{weight}{A numeric vector of weights in kilograms}
}
\value{
A numeric vector representing the calculated BMI (kg/m^2)
}
\description{
Calculates Body Mass Index (BMI) given height (in cm) and weight (in kg).
}
